# swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: Family Tree
basePath: /api
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /v1/users:
    get:
      summary: List all users
      operationId: listUsers
      tags:
        - users
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/Users'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Create a user
      operationId: createUser
      tags:
        - users
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/User'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
  User:
    type: "object"
    required:
      - id
      - login
      - name
    properties:
      id:
        type: string
        format: uuid
      login:
        type: string
      name:
        type: string
      password:
        type: string
        format: password
  Users:
    type: array
    items:
      $ref: '#/definitions/User'
...
  Error:
    type: "object"
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string